Address CheckoutController Link review
Committed-By-Agent: codex
